refactor: ErrorBoundary
