+++ apps/dashboard/app/root.tsx
@@ -1,20 +1,15 @@
 import "./app.css";
 import React from "react";
-import {
-  Meta,
-  Links,
-  Outlet,
-  Scripts,
-  ScrollRestoration,
-  isRouteErrorResponse,
-} from "react-router";
+import { Meta, Links, Outlet, Scripts, ScrollRestoration } from "react-router";
+
 import ThemeProvider from "~/providers/ThemeProvider";
 import SessionProvider from "~/providers/SessionProvider";
 
-import { meta } from "./lib/root-meta";
-import { links } from "./lib/root-links";
+import { meta } from "~/lib/root-meta";
+import { links } from "~/lib/root-links";
+import { ErrorBoundary } from "~/components/ui/ErrorBoundary";
 
-export { meta, links };
+export { meta, links, ErrorBoundary };
 
 export function Layout({ children }: { children: React.ReactNode }) {
   return (
@@ -36,33 +31,4 @@ export function Layout({ children }: { children: React.ReactNode }) {
 
 export default function App() {
   return <Outlet />;
-}
-
-export function ErrorBoundary(error: Route.ErrorBoundaryProps) {
-  let message = "Oops!";
-  let details = "An unexpected error occurred.";
-  let stack: string | undefined;
-
-  if (isRouteErrorResponse(error)) {
-    message = error.status === 404 ? "404" : "Error";
-    details =
-      error.status === 404
-        ? "The requested page could not be found."
-        : error.statusText || details;
-  } else if (import.meta.env.DEV && error && error instanceof Error) {
-    details = error.message;
-    stack = error.stack;
-  }
-
-  return (
-    <main className="pt-16 p-4 container mx-auto">
-      <h1>{message}</h1>
-      <p>{details}</p>
-      {stack && (
-        <pre className="w-full p-4 overflow-x-auto">
-          <code>{stack}</code>
-        </pre>
-      )}
-    </main>
-  );
 }
